+++ packages/mobx-react-component/test/utils.ts
@@ -18,3 +18,9 @@ export function globalSetup() {
     optimizeScheduler(unstable_batchedUpdates)
     configure({ enforceActions: "never" })
 }
+
+export function sleep(time: number) {
+    return new Promise<void>((res) => {
+        setTimeout(res, time)
+    })
+}